remove Git interface and rename GitImpl class
